backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
django-cors-headers==4.6.0
drf-spectacular==0.27.2
django-filter==24.3
django-simple-history==3.7.0
django-import-export==4.3.1
Pillow==11.0.0
psycopg[binary]==3.2.3
django-environ==0.11.2
gunicorn==23.0.0
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0
reportlab==4.2.5
pywebpush==2.0.3
cryptography==43.0.3
